testing emoji stuff. Also in the process of praising our lord and savior commands.Bot

+++ modules/budget_nitro.py
@@ -1,16 +1,31 @@
-async def emoji_converter(ctx, *text):
-    if len(text) == 0:
-        await ctx.send('You gotta put something fam.')
-        return
-
-    sender = '{} says:\n'.format(ctx.author.mention)
-    text = ' '.join(text)
-    await ctx.send(sender + text)
-    await ctx.message.delete()
+from discord.ext import commands
 
 
-def module_commands():
-    return (['nitro', 'em'], emoji_converter)
+class BudgetNitro(commands.Cog):
+    @commands.command(
+        name='nitro',
+        brief='Converts emojis tags in your message to make you feel rich',
+        description='Prints your text back out at you. Emoji codes will be ' +
+                    'converted to emojis, including animated ones and ones from ' +
+                    'other servers (hence the nitro).\n'
+                    'Type nitro list to list all available emojis.'
+    )
+    async def emoji_converter(self, ctx, *text):
+        if len(text) == 0:
+            await ctx.send('You gotta put something fam.')
+            return
+
+        if len(text) == 1 and text[0] in ['help', 'h']:
+            outstr = ''
+            for moji in ctx.bot.emojis:
+                outstr += '{}\t`:{}:`\n'.format(str(moji), moji.name)
+            await ctx.message.channel.send(outstr)
+            return
+
+        sender = '{} says:\n'.format(ctx.author.mention)
+        text = ' '.join(text)
+        await ctx.send(sender + text)
+        await ctx.message.delete()
 
     # args = msg.content.split()
     # if len(args) == 1:
